requirements.txt
boto3>=1.34
pyyaml>=6.0
requests>=2.31